chore: move cross-region-copying code to rdsutil
