+++ test/services/rds.spec.js
@@ -1,8 +1,9 @@
 // Generated by CoffeeScript 1.12.3
 (function() {
-  var AWS, helpers;
+  var AWS, helpers, rdsutil;
 
   helpers = require('../helpers');
+  rdsutil = require('../../lib/services/rdsutil');
 
   AWS = helpers.AWS;
 
@@ -50,7 +51,7 @@
           },
           paramValidation: true
         });
-        spy = helpers.spyOn(rds, 'buildCrossRegionPresignedUrl').andCallThrough();
+        spy = helpers.spyOn(rdsutil, 'buildCrossRegionPresignedUrl').andCallThrough();
         helpers.spyOn(AWS.RDS.prototype, 'getSkewCorrectedDate').andReturn(new Date(0));
         return helpers.spyOn(rds, 'getSkewCorrectedDate').andReturn(new Date(0));
       });
